Remove a:module arguments
Remove the 'a:module' arguments from the calls to `ghcmod#build_command` in
the sig, split, info, and type functions. No version check because they've
actually been optional since v5.0, but now they have been removed completely.

Also replace them with `...` in the function's argument list, so that calls
with or without them are accepted, In case there are any other plugins based
on ghcmod-vim. For the same reason, `ghcmod#detect_module` is still there.